+++ rx2-idler/src/main/java/com/squareup/rx2/idler/DelegatingIdlingResourceScheduler.java
@@ -1,6 +1,7 @@
 package com.squareup.rx2.idler;
 
-import android.support.annotation.RestrictTo;
+import androidx.annotation.Nullable;
+import androidx.annotation.RestrictTo;
 import io.reactivex.Scheduler;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.disposables.Disposable;
@@ -8,14 +9,14 @@ import io.reactivex.disposables.Disposables;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static android.support.annotation.RestrictTo.Scope.LIBRARY;
+import static androidx.annotation.RestrictTo.Scope.LIBRARY;
 
 @RestrictTo(LIBRARY)
 final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
   private final Scheduler delegate;
   private final String name;
   private final AtomicInteger work = new AtomicInteger();
-  private ResourceCallback callback;
+  @Nullable private ResourceCallback callback;
 
   DelegatingIdlingResourceScheduler(Scheduler delegate, String name) {
     this.delegate = delegate;
@@ -43,7 +44,7 @@ final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
         if (disposables.isDisposed()) {
           return Disposables.disposed();
         }
-        ScheduledWork work = createWork(action, 0L);
+        ScheduledWork work = createWork(action, 0L, 0L);
         Disposable disposable = delegateWorker.schedule(work);
         ScheduledWorkDisposable workDisposable = new ScheduledWorkDisposable(work, disposable);
         disposables.add(workDisposable);
@@ -54,7 +55,7 @@ final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
         if (disposables.isDisposed()) {
           return Disposables.disposed();
         }
-        ScheduledWork work = createWork(action, delayTime);
+        ScheduledWork work = createWork(action, delayTime, 0L);
         Disposable disposable = delegateWorker.schedule(work, delayTime, unit);
         disposables.add(disposable);
         ScheduledWorkDisposable workDisposable = new ScheduledWorkDisposable(work, disposable);
@@ -68,7 +69,7 @@ final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
         if (disposables.isDisposed()) {
           return Disposables.disposed();
         }
-        ScheduledWork work = createWork(action, initialDelay);
+        ScheduledWork work = createWork(action, initialDelay, period);
         Disposable disposable =
             delegateWorker.schedulePeriodically(work, initialDelay, period, unit);
         disposables.add(disposable);
@@ -92,12 +93,12 @@ final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
   }
 
   void stopWork() {
-    if (work.decrementAndGet() == 0) {
+    if (work.decrementAndGet() == 0 && callback != null) {
       callback.onTransitionToIdle();
     }
   }
 
-  ScheduledWork createWork(Runnable action, long delay) {
+  ScheduledWork createWork(Runnable action, long delay, long period) {
     if (action instanceof ScheduledWork) {
       // Unwrap any re-scheduled work. We want each scheduler to get its own state machine.
       action = ((ScheduledWork) action).delegate;
@@ -107,21 +108,24 @@ final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
       startWork();
     }
     int startingState = immediate ? ScheduledWork.STATE_SCHEDULED : ScheduledWork.STATE_IDLE;
-    return new ScheduledWork(action, startingState);
+    return new ScheduledWork(action, startingState, period > 0L);
   }
 
   final class ScheduledWork extends AtomicInteger implements Runnable {
     static final int STATE_IDLE = 0; // --> STATE_RUNNING, STATE_DISPOSED
     static final int STATE_SCHEDULED = 1; // --> STATE_RUNNING, STATE_DISPOSED
-    static final int STATE_RUNNING = 2; // --> STATE_COMPLETED, STATE_DISPOSED
+    static final int STATE_RUNNING = 2; // --> STATE_IDLE, STATE_COMPLETED, STATE_DISPOSED
     static final int STATE_COMPLETED = 3; // --> STATE_DISPOSED
     static final int STATE_DISPOSED = 4;
 
     final Runnable delegate;
 
-    ScheduledWork(Runnable delegate, int startingState) {
+    private final boolean isPeriodic;
+
+    ScheduledWork(Runnable delegate, int startingState, boolean isPeriodic) {
       super(startingState);
       this.delegate = delegate;
+      this.isPeriodic = isPeriodic;
     }
 
     @Override public void run() {
@@ -137,8 +141,8 @@ final class DelegatingIdlingResourceScheduler extends IdlingResourceScheduler {
               try {
                 delegate.run();
               } finally {
-                // Complete with a CAS to ensure we don't overwrite a disposed state.
-                compareAndSet(STATE_RUNNING, STATE_COMPLETED);
+                // Change state with a CAS to ensure we don't overwrite a disposed state.
+                compareAndSet(STATE_RUNNING, isPeriodic ? STATE_IDLE : STATE_COMPLETED);
                 stopWork();
               }
               return; // CAS success, we're done.