+++ rx3-idler/src/main/java/com/squareup/rx3/idler/IdlingResourceScheduler.java
@@ -1,7 +1,7 @@
-package com.squareup.rx2.idler;
+package com.squareup.rx3.idler;
 
 import androidx.test.espresso.IdlingResource;
-import io.reactivex.Scheduler;
+import io.reactivex.rxjava3.core.Scheduler;
 
 /** A RxJava {@link Scheduler} that is also an Espresso {@link IdlingResource}. */
 public abstract class IdlingResourceScheduler extends Scheduler implements IdlingResource {