+++ rx3-idler/src/main/java/com/squareup/rx3/idler/Rx3Idler.java
@@ -1,22 +1,22 @@
-package com.squareup.rx2.idler;
+package com.squareup.rx3.idler;
 
 import androidx.annotation.CheckResult;
 import androidx.annotation.NonNull;
 import androidx.test.espresso.IdlingRegistry;
 import androidx.test.espresso.IdlingResource;
-import io.reactivex.Scheduler;
-import io.reactivex.functions.Function;
-import java.util.concurrent.Callable;
+import io.reactivex.rxjava3.core.Scheduler;
+import io.reactivex.rxjava3.functions.Function;
+import io.reactivex.rxjava3.functions.Supplier;
 
 /**
  * Factory methods for connecting RxJava's {@link Scheduler} to Espresso's {@link IdlingResource}.
  * <p>
  * <pre><code>
  * RxJavaPlugins.setInitComputationSchedulerHandler(
- *     Rx2Idler.create("RxJava 2.x Computation Scheduler"));
+ *     Rx3Idler.create("RxJava 3.x Computation Scheduler"));
  * </code></pre>
  */
-public final class Rx2Idler {
+public final class Rx3Idler {
   /**
    * Returns a function which wraps the supplied {@link Scheduler} in one which notifies Espresso as
    * to whether it is currently executing work or not.
@@ -25,15 +25,13 @@ public final class Rx2Idler {
    */
   @SuppressWarnings("ConstantConditions") // Public API guarding.
   @CheckResult @NonNull
-  public static Function<Callable<Scheduler>, Scheduler> create(@NonNull final String name) {
+  public static Function<Supplier<Scheduler>, Scheduler> create(@NonNull final String name) {
     if (name == null) throw new NullPointerException("name == null");
-    return new Function<Callable<Scheduler>, Scheduler>() {
-      @Override public Scheduler apply(Callable<Scheduler> delegate) throws Exception {
-        IdlingResourceScheduler scheduler =
-            new DelegatingIdlingResourceScheduler(delegate.call(), name);
-        IdlingRegistry.getInstance().register(scheduler);
-        return scheduler;
-      }
+    return delegate -> {
+      IdlingResourceScheduler scheduler =
+          new DelegatingIdlingResourceScheduler(delegate.get(), name);
+      IdlingRegistry.getInstance().register(scheduler);
+      return scheduler;
     };
   }
 
@@ -51,7 +49,7 @@ public final class Rx2Idler {
     return new DelegatingIdlingResourceScheduler(scheduler, name);
   }
 
-  private Rx2Idler() {
+  private Rx3Idler() {
     throw new AssertionError("No instances");
   }
 }