+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 Change Log
 ==========
 
+Version 0.11.0 *(2020-05-19)*
+-----------------------------
+
+ * New: RxJava3 artifact
+ * Fix: stopWork() crashes when attempting to call onTransitionToIdle()
+
+
+Version 0.10.0 *(2019-11-22)*
+----------------------------
+
+ * Fix: Migrate to AndroidX.
+
+
+Version 0.9.1 *(2019-01-03)*
+----------------------------
+
+ * Fix: Support for periodic tasks.
+
+
 Version 0.9.0 *(2017-07-12)*
 ----------------------------